@@ -3,8 +3,13 @@
 namespace Domain\Currency\Contracts;
 
 use Domain\Currency\CurrencyCollection;
+use Domain\Currency\Enums\Bank;
+use Domain\Currency\Exceptions\FailedToConnectToBankException;
 
 interface CurrencyServiceInterface
 {
-    public function getCurrencyFromBank(): CurrencyCollection;
+    /**
+     * @throws FailedToConnectToBankException
+     */
+    public function getCurrencyFromBank(Bank $bank): CurrencyCollection;
 }
