@@ -3,18 +3,41 @@
 namespace App\Controller;
 
 use Domain\Currency\Contracts\CurrencyServiceInterface;
+use Domain\Currency\Enums\Bank;
 use Domain\Transformer\Currency\CurrencyTransformer;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
 class CurrencyController
 {
-    public function index(CurrencyServiceInterface $currencyService, CurrencyTransformer $transformer): JsonResponse
-    {
-        $currencies = $currencyService->getCurrencyFromBank();
-        $transformedCurrencies = $transformer->transformCollection($currencies->getCollection());
+    /**
+     * @var CurrencyServiceInterface
+     */
+    private $currencyService;
 
-        return (new JsonResponse())
-            ->setData($transformedCurrencies)
-            ->setStatusCode(JsonResponse::HTTP_OK);
+    /**
+     * @var CurrencyTransformer
+     */
+    private $transformer;
+
+    public function __construct(CurrencyServiceInterface $currencyService, CurrencyTransformer $transformer)
+    {
+        $this->currencyService = $currencyService;
+        $this->transformer = $transformer;
+    }
+
+    public function monoBank(): JsonResponse
+    {
+        $currencyCollection = $this->currencyService->getCurrencyFromBank(Bank::mono());
+        $transformedCurrencies = $this->transformer->transformCollection($currencyCollection->all());
+
+        return JsonResponse::create(['data' => $transformedCurrencies], JsonResponse::HTTP_OK);
+    }
+
+    public function privatBank(): JsonResponse
+    {
+        $currencyCollection = $this->currencyService->getCurrencyFromBank(Bank::privat());
+        $transformedCurrencies = $this->transformer->transformCollection($currencyCollection->all());
+
+        return JsonResponse::create(['data' => $transformedCurrencies], JsonResponse::HTTP_OK);
     }
 }
