@@ -2,7 +2,7 @@
 
 namespace App\EventListener;
 
-use Domain\Currency\Exceptions\FailedToConnectToBankException;
+use Domain\Core\Exceptions\Http\HttpException;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
@@ -10,16 +10,14 @@ use Symfony\Component\HttpKernel\KernelEvents;
 
 class ExceptionSubscriber implements EventSubscriberInterface
 {
-    use ExceptionResponses;
-
     public function onKernelException(ExceptionEvent $event)
     {
         $e = $event->getThrowable();
 
-        if ($e instanceof FailedToConnectToBankException) {
-            $response = $this->getFailedToConnectToBankExceptionResponse($e);
+        if ($e instanceof HttpException) {
+            $response = self::jsonResponse($e, $e->getStatusCode());
         } else {
-            $response = $this->defaultResponse($e, JsonResponse::HTTP_INTERNAL_SERVER_ERROR);
+            $response = self::jsonResponse($e, JsonResponse::HTTP_INTERNAL_SERVER_ERROR);
         }
 
         $event->setResponse($response);
@@ -30,5 +28,12 @@ class ExceptionSubscriber implements EventSubscriberInterface
         return [
             KernelEvents::EXCEPTION => 'onKernelException'
         ];
+    }
+
+    private static function jsonResponse(\Throwable $e, int $statusCode): JsonResponse
+    {
+        return (new JsonResponse())
+            ->setData(['message' => $e->getMessage()])
+            ->setStatusCode($statusCode);
     }
 }
